% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.rr.dim}
\alias{calc.rr.dim}
\title{Calculate the length of an input vector used to populate a rank reduced matrix for a factor analysis}
\usage{
calc.rr.dim(nj, nf)
}
\arguments{
\item{nj}{number of columns in full matrix}

\item{nf}{number of factors in reduced matrix}
}
\value{
integer dimension
}
\description{
Calculate the length of an input vector used to populate a rank reduced matrix for a factor analysis
}
