% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{cast.user}
\alias{cast.user}
\title{init.options() helper function: converts user.class to integer if not empty and not integer}
\usage{
cast.user(user)
}
\arguments{
\item{user}{Vector of classification vector set by user}
}
\value{
integer classification vector
}
\description{
init.options() helper function: converts user.class to integer if not empty and not integer
}
