% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.clustTMB}
\alias{coef.clustTMB}
\title{Get fixed-effect coefficients}
\usage{
\method{coef}{clustTMB}(object, complete = FALSE, ...)
}
\arguments{
\item{object}{The fitted clustTMB model}

\item{complete}{Currently ignored}

\item{...}{Currently ignored}
}
\value{
names numeric vector
}
\description{
Get fixed-effect coefficients
}
