% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{cormat.correction}
\alias{cormat.correction}
\title{Apply correction if NA in correlation matrix of observation subset}
\usage{
cormat.correction(cormat., ymat., nj.)
}
\arguments{
\item{cormat.}{Correlation matrix of observation subset}

\item{ymat.}{Observation subset based on cluster}

\item{nj.}{Number of response columns in observation matrix}
}
\value{
Corrected correlation matrix without NA values
}
\description{
Apply correction if NA in correlation matrix of observation subset
}
