% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{DimList}
\alias{DimList}
\title{An R7 class to generate the model dimension list}
\usage{
DimList(
  n.i = class_missing,
  n.j = class_missing,
  n.t = class_missing,
  n.k.g = class_missing,
  n.k.e = class_missing,
  n.g = class_missing,
  n.r.g = class_missing,
  n.r.e = class_missing,
  n.f.rand = class_missing,
  n.f.sp = class_missing,
  n.v = class_missing,
  nl.fix = class_missing,
  nl.rand = class_missing,
  nl.sp = class_missing
)
}
\description{
An R7 class to generate the model dimension list
}
\section{Slots}{

\describe{
\item{\code{n.i}}{Number of observations}

\item{\code{n.j}}{Number of response columns}

\item{\code{n.t}}{Number of time steps}

\item{\code{n.k.g}}{Number of covariates in gating model}

\item{\code{n.k.e}}{Number of covariates in expert model}

\item{\code{n.g}}{Number of clusters}

\item{\code{n.r.g}}{Spatial range in gating model}

\item{\code{n.r.e}}{Spatial range in expert model}

\item{\code{n.f.rand}}{Number of factors in random rank reduction}

\item{\code{n.f.sp}}{Number of factors in spatial rank reduction}

\item{\code{n.v}}{Number of vertices in spatial INLA mesh}

\item{\code{nl.fix}}{Number of loadings in covariance matrix}

\item{\code{nl.rand}}{Number of factor loading in random rank reduction}

\item{\code{nl.sp}}{Number of factor loading in spatial rank reduction}
}}

