% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extractAIC.clustTMB}
\alias{extractAIC.clustTMB}
\title{Extract the AIC of a clustTMB model}
\usage{
\method{extractAIC}{clustTMB}(object, scale, k = 2, ...)
}
\arguments{
\item{object}{The fitted clustTMB model}

\item{scale}{The scale (not used)}

\item{k}{Penalization parameter, defaults to 2}

\item{...}{Anything else}
}
\value{
numeric value
}
\description{
Extract the AIC of a clustTMB model
}
