% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_utils.R
\name{fbx}
\alias{fbx}
\title{(f)ind (b)ars e(x)tended: recursive}
\usage{
fbx(term, debug = FALSE, specials = character(0), default.special = "us")
}
\arguments{
\item{term}{a formula or piece of a formula}

\item{debug}{(logical) debug?}

\item{specials}{list of special terms}

\item{default.special}{character: special to use for parenthesized terms - i.e. random effects terms with unspecified structure
\enumerate{
\item atom (not a call or an expression): NULL
\item special, i.e. foo(...) where "foo" is in specials: return term
\item parenthesized term: \emph{if} the head of the head is | (i.e.
it is of the form (xx|gg), then convert it to the default
special type; we won't allow pathological cases like
((xx|gg))
}}
}
\description{
(f)ind (b)ars e(x)tended: recursive
}
\examples{
splitForm(quote(us(x,n=2)))
}
\keyword{internal}
