% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_utils.R
\name{findbars_x}
\alias{findbars_x}
\title{find and process random effects terms}
\usage{
findbars_x(
  term,
  debug = FALSE,
  specials = character(0),
  default.special = "norm"
)
}
\arguments{
\item{term}{a formula or piece of a formula}

\item{debug}{(logical) debug?}

\item{specials}{list of special terms}

\item{default.special}{character: special to use for parenthesized terms - i.e. random effects terms with unspecified structure
\enumerate{
\item atom (not a call or an expression): NULL
\item special, i.e. foo(...) where "foo" is in specials: return term
\item parenthesized term: \emph{if} the head of the head is | (i.e.
it is of the form (xx|gg), then convert it to the default
special type; we won't allow pathological cases like
((xx|gg))
}}
}
\value{
a list of expressions
}
\description{
find and process random effects terms
}
\examples{
splitForm(quote(us(x,n=2)))
findbars_x(~ 1 + (1 | f) + (1 | g))
findbars_x(~ 1 + (1 | f) + (1 | g))
findbars_x(~ 1 + (1|Subject))
findbars_x(~ (1||Subject))
findbars_x(~ (1|Subject))
findbars_x(~ 1 + x)
}
