% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixStruct.lookup}
\alias{fixStruct.lookup}
\title{fixStruct look-up table}
\usage{
fixStruct.lookup(fixStruct)
}
\arguments{
\item{fixStruct}{user input character vector}
}
\value{
numeric value associated with character vector
}
\description{
fixStruct look-up table
}
\examples{
fixStruct.lookup("E")
}
