% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixStruct.names}
\alias{fixStruct.names}
\title{Fixed Covariance Structure names}
\usage{
fixStruct.names()
}
\value{
character vector naming available fixed Covariance Structures
}
\description{
Fixed Covariance Structure names
}
\examples{
fixStruct.names()
}
