% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{genInitMethods}
\alias{genInitMethods}
\title{Apply classification method dependent on init.method}
\usage{
genInitMethods(n.g, n.i, n.j, control, y)
}
\arguments{
\item{n.g}{Number of clusters}

\item{n.i}{Number of observations}

\item{n.j}{Number of columns}

\item{control}{Classification settings from init.options()}

\item{y}{Observations}
}
\value{
classification vector
}
\description{
Apply classification method dependent on init.method
}
