% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.clustTMB}
\alias{logLik.clustTMB}
\title{Extract the log likelihood of a clustTMB model}
\usage{
\method{logLik}{clustTMB}(object, ...)
}
\arguments{
\item{object}{The fitted clustTMB model}

\item{...}{Currently ignored}
}
\value{
object of class \code{logLik} with attributes
}
\description{
Extract the log likelihood of a clustTMB model
}
