% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{mc.qclass}
\alias{mc.qclass}
\title{mc.qclass: quantile function from mclust. Defaults used to initiate 'E' or 'V' models when no covariates in expert/gating model}
\usage{
mc.qclass(x, k)
}
\arguments{
\item{x}{A numeric vector of observations for classification}

\item{k}{Integer specifying the number of mixtures}
}
\value{
classification vector
}
\description{
mc.qclass: quantile function from mclust. Defaults used to initiate 'E' or 'V' models when no covariates in expert/gating model
}
\keyword{internal}
