% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkRandom}
\alias{mkRandom}
\title{mkRandom: set up random effects component of the model}
\usage{
mkRandom(
  expertformula,
  gatingformula,
  expertdata,
  gatingdata,
  spatial.list,
  dim.list
)
}
\arguments{
\item{expertformula}{}

\item{gatingformula}{}

\item{expertdata}{}

\item{gatingdata}{}

\item{spatial.list}{}

\item{dim.list}{}
}
\value{

}
\description{
mkRandom: set up random effects component of the model
}
