% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{name.hc.options}
\alias{name.hc.options}
\title{init.options() helper function: assigns names to hc.options list if missing}
\usage{
name.hc.options(methods)
}
\arguments{
\item{methods}{list of hc.option methods}
}
\value{
named list of hc.option methods
}
\description{
init.options() helper function: assigns names to hc.options list if missing
}
