% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parm.lookup}
\alias{parm.lookup}
\title{Parameter Information}
\usage{
parm.lookup()
}
\value{
Description of parameters, including dimension and structure
}
\description{
Parameter Information
}
\examples{
parm.lookup()
}
