% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{reset.defaults}
\alias{reset.defaults}
\title{Reset defaults based on family, dimension, and expert/gating models}
\usage{
reset.defaults(fam, control, gate.mod, exp.mod, n.j)
}
\arguments{
\item{fam}{distribution family}

\item{control}{list of options for setting up initial classification values}

\item{gate.mod}{true if covariates in the gating model}

\item{exp.mod}{true if covariates in the expert model}
}
\value{
list of options for setting up initial classification values
}
\description{
Reset defaults based on family, dimension, and expert/gating models
}
