% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{run.mahala}
\alias{run.mahala}
\title{Updates initial class when covariates in expert formula using the Mahalanobis distance criteria}
\usage{
run.mahala(z., y., x., family, max.it = 1000)
}
\arguments{
\item{z.}{initial classification matrix}

\item{y.}{response matrix}

\item{x.}{covariates}

\item{max.it}{maximum interation}
}
\value{
updated classification matrix
}
\description{
Updates initial class when covariates in expert formula using the Mahalanobis distance criteria
}
\keyword{internal}
