% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustTMB.R
\name{run.options}
\alias{run.options}
\title{Run Options}
\usage{
run.options(check.input = NULL, run.model = NULL, do.sdreport = NULL)
}
\arguments{
\item{check.input}{TRUE: Return initial values before running TMB}

\item{run.model}{FALSE: Return TMB object before optimizing model}

\item{do.sdreport}{TRUE: Run delta method to obtain standard errors}
}
\value{
list
}
\description{
Run Options
}
\examples{
run.options()
}
