% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{set.BetaTheta}
\alias{set.BetaTheta}
\title{Set initial values for betad, theta, and thetaf}
\usage{
set.BetaTheta(Data., inits)
}
\arguments{
\item{Data.}{Initial Data list}

\item{inits}{Initial mu, var, and power}
}
\value{
List of initial values for betad, theta, and thetaf
}
\description{
Set initial values for betad, theta, and thetaf
}
