% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{set.Loadings}
\alias{set.Loadings}
\title{Set initial values for loadings parameters}
\usage{
set.Loadings(Data., cormat., corvec., dimlist.)
}
\arguments{
\item{Data.}{Initial data list}

\item{cormat.}{Correlation matrix for observation subset}

\item{corvec.}{Vector of off-diaganal correlation parameters for observation subset}

\item{dimlist.}{List of parameter dimensions}
}
\value{
List of initial loadings parameters
}
\description{
Set initial values for loadings parameters
}
