% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInit.R
\name{set.MuVarPow}
\alias{set.MuVarPow}
\title{Set initial values for mu, var, and power}
\usage{
set.MuVarPow(Class., ysub, expmod, Xd, family.)
}
\arguments{
\item{Class.}{Inital classificaiton}

\item{ysub}{Subset of observations based on column and cluster}

\item{expmod}{True if covariates in the expert model}

\item{Xd}{Expert covariates}

\item{family.}{Distribution family and link function}
}
\value{
List of initial values for mu, var, and power
}
\description{
Set initial values for mu, var, and power
}
