% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup.projDat}
\alias{setup.projDat}
\title{Setup projection data for mkDat}
\usage{
setup.projDat(mesh, projection.dat, expert.formula, gating.formula)
}
\arguments{
\item{mesh}{spatial constrained Delaunay triangulation derived from the fmesher R package}

\item{projection.dat}{Spatial Points class of projection coordinates or Spatial Points Dataframe containing projection coordinates and projection covariates}

\item{expert.formula}{Formula defining expert model. This formula corresponds to the covariates included in the response densities. Defaults to intercept only (~1) when no covariates are used.}

\item{gating.formula}{Formula defining gating model. This formula corresponds to the covariates included in the mixing proportions (logistic regression). Defaults to intercept only (~1) when no covariates are used. When a random effects term is included in the gating network, this formula will be updated so that the intercept term is removed.}
}
\value{
list of projection data
}
\description{
Setup projection data for mkDat
}
