% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{setup.spatialDat}
\alias{setup.spatialDat}
\title{Setup spatial data for mkDat}
\usage{
setup.spatialDat(n.i, spatial.list, projection.dat)
}
\arguments{
\item{n.i}{number of observations}

\item{spatial.list}{list of spatial locations and mesh}

\item{projection.dat}{Points class of projection coordinates or Spatial Points Dataframe containing projection coordinates and projection covariates}
}
\value{
list of spatial mesh and sparse A matrix
}
\description{
Setup spatial data for mkDat
}
\keyword{internal}
