% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewness}
\alias{skewness}
\title{Calculates skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{numeric vector of values for which skewness is calculated}
}
\value{
skewness value of x
}
\description{
Calculates skewness
}
