% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start.names}
\alias{start.names}
\title{Names of parameters with initial values that can be modified}
\usage{
\method{start}{names}()
}
\description{
Names of parameters with initial values that can be modified
}
