% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tweedie}
\alias{tweedie}
\title{Tweedie family and link specification}
\usage{
tweedie(link = "log")
}
\arguments{
\item{link}{link function association with family}
}
\description{
Tweedie family and link specification
}
\examples{
fam <- tweedie()
fam$family
fam$link
}
