% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extractAIC.clustTMB}
\alias{extractAIC.clustTMB}
\title{Extract the AIC of a clustTMB model}
\usage{
\method{extractAIC}{clustTMB}(fit, scale, k = 2, ...)
}
\arguments{
\item{fit}{The fitted clustTMB model}

\item{scale}{The scale, currently ignored}

\item{k}{Penalization parameter, defaults to 2}

\item{...}{Currently ignored}
}
\value{
numeric value
}
\description{
Extract the AIC of a clustTMB model
}
