% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkRandom}
\alias{mkRandom}
\title{mkRandom: set up random effects component of the model}
\usage{
mkRandom(
  expertformula,
  gatingformula,
  expertdata,
  gatingdata,
  spatial.list,
  dim.list
)
}
\arguments{
\item{expertformula}{Formula defining expert model.}

\item{gatingformula}{Formula defining gating model.}

\item{expertdata}{Data frame containing expert model covariates.}

\item{gatingdata}{Data frame containing gating model covariates.}

\item{spatial.list}{List of data objects needed when fitting a spatial GMRF model}

\item{dim.list}{List of model dimensions}
}
\value{
list vector containing random effects components of the model
}
\description{
mkRandom: set up random effects component of the model
}
