% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.clustTMB}
\alias{print.clustTMB}
\title{Print brief model summary}
\usage{
\method{print}{clustTMB}(x, ...)
}
\arguments{
\item{x}{The fitted clustTMB model}

\item{...}{Not used}
}
\value{
numeric matrix of parameter estimate and standard errors
}
\description{
Invoke TMB's print.report function
}
