% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{spdeStruct}
\alias{spdeStruct}
\title{Generate SPDE list needed as imput to FEM-SPDE spatial TMB models}
\usage{
spdeStruct(mesh)
}
\arguments{
\item{mesh}{INLA mesh object generated from inla.mesh.create or inla.mesh.2d}
}
\value{
spde List used as input into TMB anisotropic model
}
\description{
Generate SPDE list needed as imput to FEM-SPDE spatial TMB models
}
\keyword{internal}
