% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{spdeStruct}
\alias{spdeStruct}
\title{Generate SPDE list needed as input to FEM-SPDE spatial TMB models}
\usage{
spdeStruct(mesh)
}
\arguments{
\item{mesh}{mesh object generated from \code{\link[fmesher:fm_rcdt_2d]{fmesher::fm_rcdt_2d()}} or \code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d()}}}
}
\value{
spde List used as input into TMB anisotropic model
}
\description{
Generate SPDE list needed as input to FEM-SPDE spatial TMB models
}
\keyword{internal}
